backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=2.0.0
asyncpg>=0.27.0
python-dotenv>=0.19.0
httpx>=0.24.0
redis>=5.0.0
python-jose>=3.3.0
passlib>=1.7.4
pydantic>=2.0.0
email-validator>=1.1.3 